+++ acceptance-tests/fake-api/database/db.js
@@ -1,1 +1,11 @@
-export const db = {};
+import { assets } from './assets.js';
+import { workspaces } from './workspaces.js';
+import { data } from './systems.js';
+import { locations } from './locations.js';
+
+export const db = {
+    assets,
+    workspaces,
+    data,
+    locations
+};