+++ acceptance-tests/fake-api/database/systems.js
@@ -1,0 +1,5 @@
+export const data = [
+    { alias: 'System-1', id: 'SYSTEM-1' },
+    { alias: 'System-2', id: 'SYSTEM-2' },
+    { alias: 'System-3', id: 'SYSTEM-3' }
+]