+++ src/datasources/system/SystemsQueryBuilder.constants.ts
@@ -131,6 +131,12 @@ export const defaultProjection = [
 
 export const defaultOrderBy = 'createdTimeStamp DESC';
 
+export const systemFields = {
+    ID: 'id',
+    ALIAS: 'alias',
+    SCAN_CODE: 'scanCode',
+} as const;
+
 /**
  * Fields that require special boolean handling with .Equals() method
  */